Add files via upload
The Ethernet way to command the relays is finally here ! Otherwise, the Serial way also got some changes : the functions to command the doors have been fusionned in one function that require 1 parameter in order to save a few lines of code and make it clear. Also some small changes in the README.md file.